requirements.txt
boto3
picamera
playsound
RPi.gpio
gpizero
configparser
